reset auto_increment for all; add auth check; add more default values

+++ backend/src/main/java/com/keliocity/backend/controller/DatabaseFillerController.java
@@ -40,12 +40,21 @@ public class DatabaseFillerController {
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
     public String create(@RequestBody DatabaseFillerDTO dbFillerDTO) {
+        if (dbFillerDTO.getAuth() != null && !(dbFillerDTO.getAuth().equals("AUTH_PR0C0M_k3l10c1ty"))) {
+            return "not authentified";
+        }
+
         if (dbFillerDTO.getReset()) {
             meetingEmployeeRepo.deleteAll();
+            meetingEmployeeRepo.resetAutoIncrement();
             meetingRepo.deleteAll();
+            meetingRepo.resetAutoIncrement();
             employeeRepo.deleteAll();
+            employeeRepo.resetAutoIncrement();
             deskRepo.deleteAll();
+            deskRepo.resetAutoIncrement();
             roomRepo.deleteAll();
+            roomRepo.resetAutoIncrement();
             floorRepo.deleteAll();
             floorRepo.resetAutoIncrement();
         }
@@ -53,23 +62,27 @@ public class DatabaseFillerController {
         HashMap<String, RoomType> roomTypes = new HashMap<String, RoomType>();
         List<RoomType> roomTypeList = roomTypeRepo.findAll();
         for (int i=0; i<roomTypeList.size(); i++) {
-            roomTypes.put(roomTypeList.get(i).getRoomtypeName(), roomTypeList.get(i));
+            roomTypes.put(roomTypeList.get(i).getRoomtypeName().trim().toUpperCase(), roomTypeList.get(i));
         }
         HashMap<String, DeskType> deskTypes = new HashMap<String, DeskType>();
         List<DeskType> deskTypeList = deskTypeRepo.findAll();
         for (int i=0; i<deskTypeList.size(); i++) {
-            deskTypes.put(deskTypeList.get(i).getRoomType().getRoomtypeName()+"_"+deskTypeList.get(i).getDeskNumber(), deskTypeList.get(i));
+            deskTypes.put((deskTypeList.get(i).getRoomType().getRoomtypeName()+"_"+deskTypeList.get(i).getDeskNumber()).trim().toUpperCase(), deskTypeList.get(i));
         }
 
         HashMap<String, Floor> floors = new HashMap<String, Floor>();
         List<Floor> floorList = floorRepo.findAll();
         for (int i=0; i<floorList.size(); i++) {
-            floors.put(floorList.get(i).getFloorName(), floorList.get(i));
+            floors.put(floorList.get(i).getFloorName().trim().toUpperCase(), floorList.get(i));
         }
         for (int i=0; i<dbFillerDTO.getFloors().size(); i++) {
-            floors.put(dbFillerDTO.getFloors().get(i).getFloorName(), floorRepo.save(
+            String floorname = dbFillerDTO.getFloors().get(i).getFloorName();
+            if (floorname == null) {
+                floorname = "floor_"+floors.size();
+            }
+            floors.put(floorname.trim().toUpperCase(), floorRepo.save(
                 Floor.builder()
-                    .floorName(dbFillerDTO.getFloors().get(i).getFloorName())
+                    .floorName(floorname)
                     .lengthX(dbFillerDTO.getFloors().get(i).getLengthX())
                     .lengthZ(dbFillerDTO.getFloors().get(i).getLengthZ())
                     .build()
@@ -80,17 +93,26 @@ public class DatabaseFillerController {
         HashMap<String, Room> rooms = new HashMap<String, Room>();
         List<Room> roomList = roomRepo.findAll();
         for (int i=0; i<roomList.size(); i++) {
-            rooms.put(roomList.get(i).getRoomName(), roomList.get(i));
+            rooms.put(roomList.get(i).getRoomName().trim().toUpperCase(), roomList.get(i));
         }
         for (int i=0; i<dbFillerDTO.getRooms().size(); i++) {
-            rooms.put(dbFillerDTO.getRooms().get(i).getRoomName(), roomRepo.save(
+            String floorname = dbFillerDTO.getRooms().get(i).getFloorName();
+            if (dbFillerDTO.getRooms().get(i).getFloorName() == null) {
+                floorname = floors.entrySet().iterator().next().getKey();
+            }
+            floorname = floorname.trim().toUpperCase();
+            String roomname = dbFillerDTO.getRooms().get(i).getRoomName();
+            if (roomname == null) {
+                roomname = floorname+"_room_"+rooms.size();
+            }
+            rooms.put(dbFillerDTO.getRooms().get(i).getRoomName().trim().toUpperCase(), roomRepo.save(
                 Room.builder()
-                    .roomType(roomTypes.get(dbFillerDTO.getRooms().get(i).getRoomType()))
-                    .roomName(dbFillerDTO.getRooms().get(i).getRoomName())
+                    .roomType(roomTypes.get(dbFillerDTO.getRooms().get(i).getRoomType().trim().toUpperCase()))
+                    .roomName(roomname)
                     .coordX1(dbFillerDTO.getRooms().get(i).getCoordX1()).coordZ1(dbFillerDTO.getRooms().get(i).getCoordZ1())
                     .orientationDeg(dbFillerDTO.getRooms().get(i).getOrientationDeg())
                     .openspaceNumber(dbFillerDTO.getRooms().get(i).getOpenspaceNumber())
-                    .floor(floors.get(dbFillerDTO.getRooms().get(i).getFloorName()))
+                    .floor(floors.get(floorname))
                     .build()
             ));
         }
@@ -100,47 +122,59 @@ public class DatabaseFillerController {
         List<Desk> deskList = deskRepo.findAll();
         HashMap<String, String> deskNames = new HashMap<String, String>();
         for (int i=0; i<deskList.size(); i++) {
-            desks.put(deskList.get(i).getDeskName(), deskList.get(i));
-            deskNames.put(deskList.get(i).getRoom().getRoomName()+"_"+deskList.get(i).getDeskType().getDeskNumber(), deskList.get(i).getDeskName());
+            desks.put(deskList.get(i).getDeskName().trim().toUpperCase(), deskList.get(i));
+            deskNames.put((deskList.get(i).getRoom().getRoomName()+"_"+deskList.get(i).getDeskType().getDeskNumber()).trim().toUpperCase(), deskList.get(i).getDeskName());
         }
         for (int i=0; i<dbFillerDTO.getDesks().size(); i++) {
-            desks.put(dbFillerDTO.getDesks().get(i).getDeskName(), deskRepo.save(
+            Integer desknumber = dbFillerDTO.getDesks().get(i).getDeskNumber();
+            if (rooms.get(dbFillerDTO.getDesks().get(i).getRoomName().trim().toUpperCase()).getRoomType().getRoomtypeName().equals("Office1Desk") && (desknumber == null || desknumber == 0)) {
+                desknumber = 1;
+            }
+            String deskname = dbFillerDTO.getDesks().get(i).getDeskName();
+            if (deskname == null) {
+                deskname = dbFillerDTO.getDesks().get(i).getRoomName()+"_"+desknumber;
+            }
+            desks.put(deskname.trim().toUpperCase(), deskRepo.save(
                 Desk.builder()
-                    .deskName(dbFillerDTO.getDesks().get(i).getDeskName())
-                    .room(rooms.get(dbFillerDTO.getDesks().get(i).getRoomName()))
-                    .deskType(deskTypes.get(rooms.get(dbFillerDTO.getDesks().get(i).getRoomName()).getRoomType().getRoomtypeName()+"_"+dbFillerDTO.getDesks().get(i).getDeskNumber()))
+                    .deskName(deskname)
+                    .room(rooms.get(dbFillerDTO.getDesks().get(i).getRoomName().trim().toUpperCase()))
+                    .deskType(deskTypes.get((rooms.get(dbFillerDTO.getDesks().get(i).getRoomName().trim().toUpperCase()).getRoomType().getRoomtypeName()+"_"+desknumber).trim().toUpperCase()))
                     .build()
             ));
-            deskNames.put(dbFillerDTO.getDesks().get(i).getRoomName()+"_"+dbFillerDTO.getDesks().get(i).getDeskNumber(),dbFillerDTO.getDesks().get(i).getDeskName());
+            deskNames.put((dbFillerDTO.getDesks().get(i).getRoomName()+"_"+desknumber).trim().toUpperCase(),deskname.trim().toUpperCase());
         }
         deskRepo.flush();
 
         HashMap<String, Employee> employees = new HashMap<String, Employee>();
         List<Employee> employeeList = employeeRepo.findAll();
         for (int i=0; i<employeeList.size(); i++) {
-            employees.put(employeeList.get(i).getFirstName()+"_"+employeeList.get(i).getLastName(), employeeList.get(i));
+            employees.put((employeeList.get(i).getFirstName()+"_"+employeeList.get(i).getLastName()).trim().toUpperCase(), employeeList.get(i));
         }
         Desk desk = null;
         for (int i=0; i<dbFillerDTO.getEmployees().size(); i++) {
-            if (dbFillerDTO.getEmployees().get(i).getRoomName() != null && dbFillerDTO.getEmployees().get(i).getDeskNumber() != null) {
-                desk = desks.get(deskNames.get(dbFillerDTO.getEmployees().get(i).getRoomName()+"_"+dbFillerDTO.getEmployees().get(i).getDeskNumber()));
+            Integer desknumber = dbFillerDTO.getEmployees().get(i).getDeskNumber();
+            if (rooms.get(dbFillerDTO.getEmployees().get(i).getRoomName().trim().toUpperCase()).getRoomType().getRoomtypeName().equals("Office1Desk") && (desknumber == null || desknumber == 0)) {
+                desknumber = 1;
+            }
+            if (dbFillerDTO.getEmployees().get(i).getRoomName() != null && desknumber != null) {
+                desk = desks.get(deskNames.get((dbFillerDTO.getEmployees().get(i).getRoomName()+"_"+dbFillerDTO.getEmployees().get(i).getDeskNumber()).trim().toUpperCase()));
                 if (desk == null) {
-                    String deskname = dbFillerDTO.getEmployees().get(i).getRoomName()+"_"+dbFillerDTO.getEmployees().get(i).getDeskNumber();
+                    String deskname = (dbFillerDTO.getEmployees().get(i).getRoomName()+"_"+desknumber);
                     if (dbFillerDTO.getEmployees().get(i).getDeskName() != null) {
                         deskname = dbFillerDTO.getEmployees().get(i).getDeskName();
                     }
                     desk = deskRepo.save(
                         Desk.builder()
                             .deskName(deskname)
-                            .room(rooms.get(dbFillerDTO.getEmployees().get(i).getRoomName()))
-                            .deskType(deskTypes.get(rooms.get(dbFillerDTO.getEmployees().get(i).getRoomName()).getRoomType().getRoomtypeName()+"_"+dbFillerDTO.getEmployees().get(i).getDeskNumber()))
+                            .room(rooms.get(dbFillerDTO.getEmployees().get(i).getRoomName().trim().toUpperCase()))
+                            .deskType(deskTypes.get((rooms.get(dbFillerDTO.getEmployees().get(i).getRoomName().trim().toUpperCase()).getRoomType().getRoomtypeName()+"_"+desknumber).trim().toUpperCase()))
                             .build()
                     );
-                    desks.put(deskname, desk);
-                    deskNames.put(dbFillerDTO.getEmployees().get(i).getRoomName()+"_"+dbFillerDTO.getEmployees().get(i).getDeskNumber(),deskname);
+                    desks.put(deskname.trim().toUpperCase(), desk);
+                    deskNames.put((dbFillerDTO.getEmployees().get(i).getRoomName()+"_"+desknumber).trim().toUpperCase(),deskname);
                 }
             } else if (dbFillerDTO.getEmployees().get(i).getDeskName() != null) {
-                desk = desks.get(dbFillerDTO.getEmployees().get(i).getDeskName());
+                desk = desks.get(dbFillerDTO.getEmployees().get(i).getDeskName().trim().toUpperCase());
             }
 
             String inoffice = dbFillerDTO.getEmployees().get(i).getInOffice();
@@ -155,7 +189,7 @@ public class DatabaseFillerController {
             if (dbFillerDTO.getEmployees().get(i).getSprite() == null) {
                 sprite = "MAN1";
             }
-            employees.put(dbFillerDTO.getEmployees().get(i).getFirstName()+"_"+dbFillerDTO.getEmployees().get(i).getLastName(), employeeRepo.save(
+            employees.put((dbFillerDTO.getEmployees().get(i).getFirstName()+"_"+dbFillerDTO.getEmployees().get(i).getLastName()).trim().toUpperCase(), employeeRepo.save(
                     Employee.builder()
                         .firstName(dbFillerDTO.getEmployees().get(i).getFirstName())
                         .lastName(dbFillerDTO.getEmployees().get(i).getLastName())
@@ -174,16 +208,23 @@ public class DatabaseFillerController {
         HashMap<String, Meeting> meetings = new HashMap<String, Meeting>();
         List<Meeting> meetingList = meetingRepo.findAll();
         for (int i=0; i<meetingList.size(); i++) {
-            meetings.put(meetingList.get(i).getTitle(), meetingList.get(i));
+            meetings.put(meetingList.get(i).getTitle().trim().toUpperCase(), meetingList.get(i));
         }
         for (int i=0; i<dbFillerDTO.getMeetings().size(); i++) {
-            String deskname = dbFillerDTO.getMeetings().get(i).getDeskName();
-            if (deskname == null && dbFillerDTO.getMeetings().get(i).getDeskNumber() != null) {
-                deskname = deskNames.get(dbFillerDTO.getMeetings().get(i).getRoomName()+"_"+dbFillerDTO.getMeetings().get(i).getDeskNumber());
+            Integer desknumber = dbFillerDTO.getMeetings().get(i).getDeskNumber();
+            if (rooms.get(dbFillerDTO.getMeetings().get(i).getRoomName().trim().toUpperCase()).getRoomType().getRoomtypeName().equals("Office1Desk") && desknumber == null) {
+                desknumber = 1;
             }
-            meetings.put(dbFillerDTO.getMeetings().get(i).getTitle(), meetingRepo.save(
+            String deskname = dbFillerDTO.getMeetings().get(i).getDeskName();
+            if (deskname == null && desknumber != null) {
+                deskname = deskNames.get((dbFillerDTO.getMeetings().get(i).getRoomName()+"_"+desknumber).trim().toUpperCase());
+            }
+            if (deskname != null) {
+                deskname = deskname.trim().toUpperCase();
+            }
+            meetings.put(dbFillerDTO.getMeetings().get(i).getTitle().trim().toUpperCase(), meetingRepo.save(
                     Meeting.builder()
-                        .room(rooms.get(dbFillerDTO.getMeetings().get(i).getRoomName()))
+                        .room(rooms.get(dbFillerDTO.getMeetings().get(i).getRoomName().trim().toUpperCase()))
                         .desk(desks.get(deskname))
                         .title(dbFillerDTO.getMeetings().get(i).getTitle())
                         .startingHour(dbFillerDTO.getMeetings().get(i).getStartingHour()) // format YYYY-MM-DDTHH:mm:ss (ex: 2026-02-12T21:31:00)
@@ -197,13 +238,13 @@ public class DatabaseFillerController {
         HashMap<String, MeetingEmployee> meetingEmployees = new HashMap<String, MeetingEmployee>();
         List<MeetingEmployee> meetingEmployeeList = meetingEmployeeRepo.findAll();
         for (int i=0; i<meetingEmployeeList.size(); i++) {
-            meetingEmployees.put(meetingEmployeeList.get(i).getEmployee().getFirstName()+"_"+meetingEmployeeList.get(i).getEmployee().getLastName()+"_"+meetingEmployeeList.get(i).getMeeting().getTitle(), meetingEmployeeList.get(i));
+            meetingEmployees.put((meetingEmployeeList.get(i).getEmployee().getFirstName()+"_"+meetingEmployeeList.get(i).getEmployee().getLastName()+"_"+meetingEmployeeList.get(i).getMeeting().getTitle()).trim().toUpperCase(), meetingEmployeeList.get(i));
         }
         for (int i=0; i<dbFillerDTO.getMeetingEmployees().size(); i++) {
-            meetingEmployees.put(dbFillerDTO.getMeetingEmployees().get(i).getEmployeeFirstName()+"_"+dbFillerDTO.getMeetingEmployees().get(i).getEmployeeLastName()+"_"+dbFillerDTO.getMeetingEmployees().get(i).getMeetingTitle(), meetingEmployeeRepo.save(
+            meetingEmployees.put((dbFillerDTO.getMeetingEmployees().get(i).getEmployeeFirstName()+"_"+dbFillerDTO.getMeetingEmployees().get(i).getEmployeeLastName()+"_"+dbFillerDTO.getMeetingEmployees().get(i).getMeetingTitle()).trim().toUpperCase(), meetingEmployeeRepo.save(
                     new MeetingEmployee(
-                        meetings.get(dbFillerDTO.getMeetingEmployees().get(i).getMeetingTitle()),
-                        employees.get(dbFillerDTO.getMeetingEmployees().get(i).getEmployeeFirstName()+"_"+dbFillerDTO.getMeetingEmployees().get(i).getEmployeeLastName()),
+                        meetings.get(dbFillerDTO.getMeetingEmployees().get(i).getMeetingTitle().trim().toUpperCase()),
+                        employees.get((dbFillerDTO.getMeetingEmployees().get(i).getEmployeeFirstName()+"_"+dbFillerDTO.getMeetingEmployees().get(i).getEmployeeLastName()).trim().toUpperCase()),
                         dbFillerDTO.getMeetingEmployees().get(i).getPresent(),
                         dbFillerDTO.getMeetingEmployees().get(i).getRemote()
                 )));
